refactor(utxo-lib): improve `network` exports
For historical reason, `networks.ts` exported the values via
`exports = networks`, resulting in typescript declaration of
`networks: any` for `index.d.ts`.

There was no good reason for this behavior.

We now export `networks` as a regular value from `index.ts`, along
with a set of useful methods.

For unclear reasons there were problems when the network utility
methods and the network list were in separate files, so I decided to
put everything in a single file.

BREAKING CHANGE:
* The namespace `utxolib.coins` is removed

Issue: BG-40432
