+++ modules/utxo-lib/test/fixtures_thirdparty/download.ts
@@ -1,8 +1,7 @@
 import * as fs from 'fs-extra';
 import * as childProcess from 'child_process';
 
-import { Network } from '../../src/networkTypes';
-import { getNetworkList, getNetworkName, isMainnet } from '../../src/coins';
+import { Network, getNetworkList, getNetworkName, isMainnet } from '../../src';
 import { getArchiveUrl, getFixtureInfo, getArchiveRoot, sigHashTestFile, txValidTestFile } from './fixtures';
 
 function downloadAndUnpackTestFixtures(network: Network) {
