+++ modules/utxo-lib/test/fixtures_thirdparty/fixtures.ts
@@ -1,7 +1,5 @@
 import * as fs from 'fs-extra';
-import { Network } from '../../src/networkTypes';
-import { getMainnet, getNetworkName } from '../../src/coins';
-import * as networks from '../../src/networks';
+import { Network, networks, getMainnet, getNetworkName } from '../../src/networks';
 
 export type FixtureInfo = {
   projectPath: string;
