+++ modules/utxo-lib/test/networks.ts
@@ -1,64 +1,80 @@
 import * as assert from 'assert';
-import { coins, Network, networks, supportsSegwit, supportsTaproot } from '../src';
+import {
+  getMainnet,
+  getNetworkList,
+  getNetworkName,
+  getTestnet,
+  isBitcoin,
+  isBitcoinCash,
+  isBitcoinGold,
+  isBitcoinSV,
+  isLitecoin,
+  isMainnet,
+  isSameCoin,
+  isTestnet,
+  isValidNetwork,
+  isZcash,
+  Network,
+  networks,
+  supportsSegwit,
+  supportsTaproot,
+} from '../src/networks';
 
 describe('networks', function () {
   // Ideally, all properties for all coins should be distinct.
   // However, there are some exceptions and some networks share the same properties.
 
   // Here we define some groups of networks that are allowed to share properties.
-  const bitcoinSharedMessagePrefix = (network) =>
-    coins.isBitcoin(network) || coins.isBitcoinCash(network) || coins.isBitcoinSV(network);
+  const bitcoinSharedMessagePrefix = (network) => isBitcoin(network) || isBitcoinCash(network) || isBitcoinSV(network);
 
   const bitcoinMainnetSharedPubkeyPrefix = (network) =>
-    coins.isMainnet(network) &&
-    (coins.isBitcoin(network) || coins.isBitcoinCash(network) || coins.isBitcoinSV(network));
+    isMainnet(network) && (isBitcoin(network) || isBitcoinCash(network) || isBitcoinSV(network));
 
   const bitcoinMainnetSharedScriptPrefix = (network) => bitcoinMainnetSharedPubkeyPrefix(network);
 
   const bitcoinTestnetSharedPubkeyPrefix = (network) =>
-    coins.isTestnet(network) &&
-    (coins.isBitcoin(network) ||
-      coins.isBitcoinCash(network) ||
-      coins.isBitcoinSV(network) ||
-      coins.isBitcoinGold(network) ||
-      coins.isLitecoin(network));
+    isTestnet(network) &&
+    (isBitcoin(network) ||
+      isBitcoinCash(network) ||
+      isBitcoinSV(network) ||
+      isBitcoinGold(network) ||
+      isLitecoin(network));
 
   const bitcoinTestnetSharedScriptPrefix = (network) =>
-    bitcoinTestnetSharedPubkeyPrefix(network) && !coins.isLitecoin(network);
+    bitcoinTestnetSharedPubkeyPrefix(network) && !isLitecoin(network);
 
   const bitcoinMainnetSharedWIFPrefix = (network) =>
-    coins.isMainnet(network) &&
-    (coins.isBitcoin(network) ||
-      coins.isBitcoinCash(network) ||
-      coins.isBitcoinGold(network) ||
-      coins.isBitcoinSV(network) ||
-      coins.isZcash(network));
+    isMainnet(network) &&
+    (isBitcoin(network) ||
+      isBitcoinCash(network) ||
+      isBitcoinGold(network) ||
+      isBitcoinSV(network) ||
+      isZcash(network));
 
-  const bech32Coins = (network) =>
-    coins.isBitcoin(network) || coins.isBitcoinGold(network) || coins.isLitecoin(network);
+  const bech32Coins = (network) => isBitcoin(network) || isBitcoinGold(network) || isLitecoin(network);
 
   const sameGroup = (group, network, otherNetwork) => group(network) && group(otherNetwork);
 
   describe('getNetworkList()', function () {
     it('mainnets are sorted alphabetically', function () {
-      const mainnets = coins.getNetworkList().filter(coins.isMainnet);
+      const mainnets = getNetworkList().filter(isMainnet);
       const sortedMainnets = [...mainnets].sort((a, b) =>
-        (coins.getNetworkName(a) as string).localeCompare(coins.getNetworkName(b) as string)
+        (getNetworkName(a) as string).localeCompare(getNetworkName(b) as string)
       );
       assert.deepStrictEqual(mainnets, sortedMainnets);
     });
 
     it('testnet(s) follow mainnets', function () {
-      const list = coins.getNetworkList();
+      const list = getNetworkList();
       while (list.length > 0) {
         // first element is a mainnet
         const mainnet = list.shift();
         assert.strict(mainnet);
-        assert.strictEqual(coins.isMainnet(mainnet), true);
+        assert.strictEqual(isMainnet(mainnet), true);
 
         // subsequent entries are testnets
-        while (list.length > 0 && coins.isTestnet(list[0])) {
-          assert.strictEqual(coins.getMainnet(list[0]), mainnet);
+        while (list.length > 0 && isTestnet(list[0])) {
+          assert.strictEqual(getMainnet(list[0]), mainnet);
           list.shift();
         }
       }
@@ -68,7 +84,7 @@ describe('networks', function () {
   describe('Features', function () {
     it('have expected values for networks', function () {
       assert.deepStrictEqual(
-        coins.getNetworkList().map((n) => [coins.getNetworkName(n), supportsSegwit(n), supportsTaproot(n)]),
+        getNetworkList().map((n) => [getNetworkName(n), supportsSegwit(n), supportsTaproot(n)]),
         [
           ['bitcoin', true, true],
           ['testnet', true, true],
@@ -94,37 +110,37 @@ describe('networks', function () {
 
     describe(`networks.${name}`, function () {
       it('is valid network', function () {
-        assert(coins.isValidNetwork(network));
+        assert(isValidNetwork(network));
       });
 
       it('getNetworkName() returns network name', function () {
-        assert.strictEqual(name, coins.getNetworkName(network));
+        assert.strictEqual(name, getNetworkName(network));
       });
 
       it('has corresponding testnet/mainnet', function () {
-        if (coins.isMainnet(network)) {
-          assert.strictEqual(coins.isTestnet(network), false);
-          assert.strictEqual(coins.getMainnet(network), network);
-          assert.strictEqual(typeof coins.getTestnet(network), 'object');
+        if (isMainnet(network)) {
+          assert.strictEqual(isTestnet(network), false);
+          assert.strictEqual(getMainnet(network), network);
+          assert.strictEqual(typeof getTestnet(network), 'object');
         } else {
-          assert.strictEqual(coins.isMainnet(network), false);
-          assert.strictEqual(coins.getTestnet(network), network);
-          assert.notStrictEqual(coins.getMainnet(network), network);
-          assert.strictEqual(typeof coins.getMainnet(network), 'object');
+          assert.strictEqual(isMainnet(network), false);
+          assert.strictEqual(getTestnet(network), network);
+          assert.notStrictEqual(getMainnet(network), network);
+          assert.strictEqual(typeof getMainnet(network), 'object');
         }
       });
 
       it('has expected properties', function () {
         assert.strictEqual(typeof network, 'object');
         assert.strictEqual(typeof network.messagePrefix, 'string');
-        assert.strictEqual(typeof network.bech32, bech32Coins(network) ? 'string' : 'undefined');
+        assert.strictEqual(typeof (network as any).bech32, bech32Coins(network) ? 'string' : 'undefined');
         assert.strictEqual(typeof network.bip32, 'object');
         assert.strictEqual(typeof network.pubKeyHash, 'number');
         assert.strictEqual(typeof network.scriptHash, 'number');
         assert.strictEqual(typeof network.wif, 'number');
         assert.strictEqual(typeof network.coin, 'string');
 
-        if (coins.isMainnet(network)) {
+        if (isMainnet(network)) {
           assert.strictEqual(network.bip32.public, networks.bitcoin.bip32.public);
           assert.strictEqual(network.bip32.private, networks.bitcoin.bip32.private);
         } else {
@@ -138,8 +154,8 @@ describe('networks', function () {
 
         it('isSameCoin() returns true testnet/mainnet variants', function () {
           assert.strictEqual(
-            coins.isSameCoin(network, otherNetwork),
-            otherNetwork === coins.getMainnet(network) || otherNetwork === coins.getTestnet(network)
+            isSameCoin(network, otherNetwork),
+            otherNetwork === getMainnet(network) || otherNetwork === getTestnet(network)
           );
 
           assert.strictEqual(name === otherName, network === otherNetwork);
@@ -152,7 +168,7 @@ describe('networks', function () {
         it(`has distinct properties with ${otherName}`, function () {
           assert.strictEqual(
             network.messagePrefix === otherNetwork.messagePrefix,
-            coins.isSameCoin(network, otherNetwork) || sameGroup(bitcoinSharedMessagePrefix, network, otherNetwork)
+            isSameCoin(network, otherNetwork) || sameGroup(bitcoinSharedMessagePrefix, network, otherNetwork)
           );
 
           assert.strictEqual(
@@ -170,7 +186,7 @@ describe('networks', function () {
           assert.strictEqual(
             network.wif === otherNetwork.wif,
             sameGroup(bitcoinMainnetSharedWIFPrefix, network, otherNetwork) ||
-              sameGroup(coins.isTestnet, network, otherNetwork)
+              sameGroup(isTestnet, network, otherNetwork)
           );
         });
       }
